refactor(sdk, client, playground): agent class moved to stores
