+++ libs/sdk/src/api/index.ts
@@ -1,6 +1,4 @@
 export * from "./agent";
-export * from "./agent-subscription";
-export * from "./agent-user-input";
 export * from "./app";
 export * from "./auth";
 export * from "./base";
