+++ libs/sdk/src/index.ts
@@ -3,6 +3,17 @@ import "./listener";
 export * from "./api";
 export * from "./constants";
 export * from "./env";
-export { createRoom, InsightStore as Insight, RoomStore } from "./stores";
+export {
+	AgentStore,
+	createRoom,
+	InsightStore as Insight,
+	isRequestUserInputAction,
+	normalizeUserInputQuestion,
+	parseUserInputRequest,
+	RoomStore,
+	type UserInputOption,
+	type UserInputQuestion,
+	type UserInputRequest,
+} from "./stores";
 export * from "./types";
 export * from "./utility";
