+++ libs/sdk/src/stores/agent/agent-user-input.spec.ts
