+++ libs/sdk/src/stores/index.ts
@@ -1,2 +1,3 @@
+export * from "./agent";
 export * from "./insight";
 export * from "./room";
